Todolist major fixes.
