+++ App.js
@@ -28,7 +28,7 @@ import DashBoard from './src/components/dashboard/DashBoard';
 import ApproveCircle from './src/components/circle/approveCircle/ApproveCircle';
 import CreateCircle from './src/components/circle/createCircle/CreateCircle';
 import InviteToCircle from './src/components/circle/inviteToCircle/InviteToCircle';
-import TodoList from './src/components/todoList/TodoList';
+import TodoList from './src/components/plugins/todoList/TodoList';
 import ChoosePlugins from './src/components/managePlugin/choosePlugin/ChoosePlugin';
 import ExpenseManager from './src/components/plugins/expenseManager/ExpenseManager';
 import Chat from './src/components/plugins/chat/Chat';
@@ -92,6 +92,14 @@ const App = () => {
                 }}
               />
               <Stack.Screen
+                name="TodoList"
+                options={{
+                  title: 'Todo List',
+                  headerLeft: null,
+                }}
+                component={TodoList}
+              />
+              <Stack.Screen
                 name="Chat"
                 component={Chat}
                 options={{
@@ -147,8 +155,6 @@ const App = () => {
             component={Bootstrap}
             options={{headerShown: false}}
           />
-          <Stack.Screen name="DashBoard" component={DashBoard} />
-          {/* <Stack.Screen name="To-do list" component={TodoList} /> */}
         </Stack.Navigator>
       </NavigationContainer>
     </Root>
