+++ src/components/dashboard/DashBoard.js
@@ -24,6 +24,10 @@ const DashBoard = ({route, navigation}) => {
   const handleChatButtonClick = async () => {
     navigation.navigate('Chat', {userInfo: bootstrapState.user});
   };
+
+  const handleTodoListButtonClick = async () => {
+    navigation.navigate('TodoList', {userInfo: bootstrapState.user});
+  };
   const handleLogoutButtonClick = async () => {
     try {
       //If user authenticate with google oAuth
@@ -87,6 +91,9 @@ const DashBoard = ({route, navigation}) => {
         </Button>
         <Button onPress={handleChatButtonClick}>
           <Text> Chat </Text>
+        </Button>
+        <Button onPress={handleTodoListButtonClick}>
+          <Text> TodoList </Text>
         </Button>
       </Content>
     </Container>
